Fix Timezone Issue
The start date for Costcodle was set to September 21, 12AM in the local timezone. This resulted in the GameNumber changing dependent on the local timezone. Ergo, UTC-5 would be on game 494, and UTC+5 would be on game 495, rather than both being on the same game number. By setting the costcodleStartDate to October 21 in UTC+0, this issue is avoided and all users should experience the same game number regardless of local timezone.